Add support to collect HAR data from all visited pages
When the flow goes through multiple pages it keeps data for all of them, not only the last visited page.
This allow retrieve data for more complex use cases like, enter on the site, do a search and check the result.
It also add a `resetHAR` function to allow clear the collected data anytime.